# lib/api-spec/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
  description: WorkforceOS - Multilingual Workforce Management Platform API
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterRequest:
      type: object
      required: [email, password, fullName, company]
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        company:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/AuthUser'
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
        companyId:
          type: integer

    # Dashboard
    DashboardStats:
      type: object
      properties:
        totalEmployees:
          type: integer
        presentToday:
          type: integer
        onLeave:
          type: integer
        pendingRequests:
          type: integer
        monthlyPayroll:
          type: number
        recentAttendance:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              present:
                type: integer
              absent:
                type: integer
        departmentStats:
          type: array
          items:
            type: object
            properties:
              department:
                type: string
              count:
                type: integer

    # Employee
...
    Department:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    DepartmentInput:
      type: object
      required: [name, companyId]
      properties:
        name:
          type: string
        companyId:
          type: integer
        description:
          type: string

    # Attendance
    Attendance:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
          nullable: true
        date:
          type: string
        clockIn:
          type: string
          nullable: true
        clockOut:
          type: string
          nullable: true
        totalHours:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        method:
          type: string
          nullable: true
        status:
          type: string
          enum: [present, absent, late, half-day]
        isLate:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    ClockInInput:
      type: object
      required: [employeeId]
      properties:
        employeeId:
          type: integer
        location:
          type: string
        method:
          type: string
    ClockOutInput:
      type: object
      required: [employeeId]
      properties:
        employeeId:
          type: integer

    # Leave
...
    Asset:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        type:
          type: string
          nullable: true
        serialNumber:
          type: string
          nullable: true
        assignedTo:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        status:
          type: string
          enum: [available, assigned, maintenance, retired]
        purchaseDate:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    AssetInput:
      type: object
      required: [companyId, name]
      properties:
        companyId:
          type: integer
        name:
          type: string
        type:
          type: string
        serialNumber:
          type: string
        assignedTo:
          type: integer
        status:
          type: string
          enum: [available, assigned, maintenance, retired]
        purchaseDate:
          type: string
    AssetUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        serialNumber:
          type: string
        assignedTo:
          type: integer
        status:
          type: string
          enum: [available, assigned, maintenance, retired]
        purchaseDate:
          type: string

    # Error
    Error:
      type: object
      properties:
        error:
          type: string

    # List responses
...
    AttendanceList:
      type: object
      properties:
        attendance:
          type: array
          items:
            $ref: '#/components/schemas/Attendance'
        total:
          type: integer
...
    AssetList:
      type: object
      properties:
        assets:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
    DepartmentList:
      type: object
      properties:
        departments:
          type: array
          items:
            $ref: '#/components/schemas/Department'
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      operationId: login
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/register:
    post:
      operationId: register
      summary: Register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      summary: Get dashboard statistics
      parameters:
        - name: companyId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'

  # Employees
...
  /departments:
    get:
      operationId: getDepartments
      summary: List departments
      parameters:
        - name: companyId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentList'
    post:
      operationId: createDepartment
      summary: Create department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'

  # Attendance
  /attendance:
    get:
      operationId: getAttendance
      summary: List attendance records
      parameters:
        - name: companyId
          in: query
          required: true
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceList'
  /attendance/clock-in:
    post:
      operationId: clockIn
      summary: Clock in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockInInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attendance'
  /attendance/clock-out:
    post:
      operationId: clockOut
      summary: Clock out
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockOutInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attendance'
  /attendance/today:
    get:
      operationId: getTodayAttendance
      summary: Get today attendance for an employee
      parameters:
        - name: employeeId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attendance'

  # Leaves
...
  /assets:
    get:
      operationId: getAssets
      summary: List assets
      parameters:
        - name: companyId
          in: query
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetList'
    post:
      operationId: createAsset
      summary: Create asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
  /assets/{id}:
    put:
      operationId: updateAsset
      summary: Update asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
    delete:
      operationId: deleteAsset
      summary: Delete asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No content