# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * WorkforceOS - Multilingual Workforce Management Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { AssetStatus } from './assetStatus';

export interface Asset {
  id?: number;
  companyId?: number;
  name?: string;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  /** @nullable */
  assignedTo?: number | null;
  /** @nullable */
  assignedToName?: string | null;
  status?: AssetStatus;
  /** @nullable */
  purchaseDate?: string | null;
  createdAt?: Date;
}