# lib/api-zod/src/generated/types/clockOutInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * WorkforceOS - Multilingual Workforce Management Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface ClockOutInput {
  employeeId: number;
}